requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch
--find-links https://download.pytorch.org/whl/torch_stable.html
torchvision
mysql-connector-python